Reconcile DB schema with PROJECT_DOCUMENTATION §3
Bring migrations, models and seeders in line with the updated docs.

Catalog:
- products: unified item type (trading|manufactured|raw|service) + capability
  flags (is_stock_tracked/purchasable/manufacturable/sellable/web_listed),
  manufacture_mode, variant_mode, markup_percent, specifications json; fix the
  type/variant_mode conflation; add the web-catalog composite index
- product_variants: rename price->retail_price, cost_price->cost; add
  wholesale_price, price_is_manual, reserved_quantity; stock as decimal(12,3)
- categories.markup_percent

Missing tables added (with models):
- customers (the buyer record for all channels; wires orders.customer_id)
- quotations + quotation_items (quote -> convert to sale; wires orders.quotation_id)
- boms + bom_items, production_orders + production_consumptions (manufacturing)

Sales/supply alignment:
- orders: channel (web|pos|vendor), customer_id, quotation_id, price_tier,
  paid_total, created_by; money decimal(15,2)
- purchases: purchase_number/reference/purchase_date/subtotal/tax_total/
  grand_total/paid_total; purchase_items qty decimal + timestamps
- stock_movements: signed decimal quantity + balance_after + unit_cost, doc type set
- suppliers: company/tax_number/opening_balance/notes
- payments.received_by; order_items qty decimal; cart_items qty decimal

Seeders:
- RolePermissionSeeder: add boms/production/stock/quotations/pos/gallery/users
  groups + procurement/production-manager/inventory-manager/cashier/sales-rep
  roles (95 permissions, 12 roles)
- CatalogSeeder: demo brands, categories, 12 web-listed products + default
  variants + reviews + walk-in customer

Verified: migrate:fresh --seed runs clean (43 migrations, 4 seeders); storefront
+ login still render; admin login works.

+++ database/migrations/2026_06_22_150013_create_carts_table.php
@@ -22,7 +22,7 @@ return new class extends Migration
             $table->id();
             $table->foreignId('cart_id')->constrained('carts')->cascadeOnDelete();
             $table->foreignId('product_variant_id')->constrained('product_variants')->cascadeOnDelete();
-            $table->integer('quantity')->default(1);
+            $table->decimal('quantity', 12, 3)->default(1);
             $table->decimal('price_snapshot', 12, 2);
             $table->timestamps();
